Fixed edge use-cases with invalid content-length and transfer-encoding.

+++ src/vmod_soap.h
@@ -87,7 +87,7 @@ typedef struct priv_soap_task {
 	struct soap_req_http		*req_http;
 	struct soap_req_xml		*req_xml;
 	int				state;
-	int				bytes_left;
+	ssize_t				bytes_left;
 } sess_record;
 
 // Common HTTP Headers
